Create FriendsList as separate component
